third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/register-bypassing-scheme-partial-expected.txt

CONSOLE ERROR: Loading the image 'http://127.0.0.1:8000/security/resources/abe.png' violates the following Content Security Policy directive: "default-src https:". Note that 'img-src' was not explicitly set, so 'default-src' is used as a fallback. The action has been blocked.
CONSOLE MESSAGE: PASS (1/12)
CONSOLE ERROR: Loading the stylesheet 'http://127.0.0.1:8000/security/resources/cssStyle.css' violates the following Content Security Policy directive: "default-src https:". Note that 'style-src-elem' was not explicitly set, so 'default-src' is used as a fallback. The action has been blocked.
CONSOLE MESSAGE: PASS (2/12)
CONSOLE MESSAGE: PASS (3/12)
CONSOLE MESSAGE: PASS (4/12)
CONSOLE ERROR: Loading the image 'http://127.0.0.1:8000/security/resources/abe.png' violates the following Content Security Policy directive: "default-src https:". Note that 'img-src' was not explicitly set, so 'default-src' is used as a fallback. The action has been blocked.
CONSOLE MESSAGE: PASS (5/12)
CONSOLE ERROR: Loading the stylesheet 'http://127.0.0.1:8000/security/resources/cssStyle.css' violates the following Content Security Policy directive: "default-src https:". Note that 'style-src-elem' was not explicitly set, so 'default-src' is used as a fallback. The action has been blocked.
CONSOLE MESSAGE: PASS (6/12)
CONSOLE MESSAGE: PASS (7/12)
CONSOLE ERROR: Loading the stylesheet 'http://127.0.0.1:8000/security/resources/cssStyle.css' violates the following Content Security Policy directive: "default-src https:". Note that 'style-src-elem' was not explicitly set, so 'default-src' is used as a fallback. The action has been blocked.
CONSOLE MESSAGE: PASS (8/12)
CONSOLE ERROR: Loading the image 'http://127.0.0.1:8000/security/resources/abe.png' violates the following Content Security Policy directive: "default-src https:". Note that 'img-src' was not explicitly set, so 'default-src' is used as a fallback. The action has been blocked.
CONSOLE MESSAGE: PASS (9/12)
CONSOLE MESSAGE: PASS (10/12)
CONSOLE ERROR: Loading the image 'http://127.0.0.1:8000/security/resources/abe.png' violates the following Content Security Policy directive: "default-src https:". Note that 'img-src' was not explicitly set, so 'default-src' is used as a fallback. The action has been blocked.
CONSOLE MESSAGE: PASS (11/12)
CONSOLE ERROR: Loading the stylesheet 'http://127.0.0.1:8000/security/resources/cssStyle.css' violates the following Content Security Policy directive: "default-src https:". Note that 'style-src-elem' was not explicitly set, so 'default-src' is used as a fallback. The action has been blocked.
CONSOLE MESSAGE: PASS (12/12)
This test ensures that registering a scheme as bypassing CSP actually bypasses CSP. This test passes if only PASSes are generated.
